pyproject.toml
[project]
name = "r-bookly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asgiref>=3.11.1",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi-mail>=1.6.2",
    "fastapi[standard]>=0.129.0",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "pyjwt>=2.11.0",
    "redis>=7.2.0",
    "sqlmodel>=0.0.34",
]
